Refactor Headers, move to Internal\Headers
